hangman: stage 5
Checking whether players enter exactly one letter. If not, print "Please, input a single letter.";

Making sure that the player entered a lowercase English letter. If not, the program should print " Please, enter a lowercase letter from the English alphabet." Remember that when players input nothing or non-letter characters, it shouldn't count as a correct input either;

If users enter the same letter twice (doesn't matter whether it appears in the word or not), then the program outputs "You've already guessed this letter.". Do not decrease the number of attempts in this case;