"Coffee machine" project completed
Write a program that puts some basic information on the screen. Let the machine describe what it takes to make a cup of coffee!
The machine should calculate how many ingredients it needs to make the required amount of coffee.
Program the machine to calculate how many cups it can make based on the number of set ingredients.
Teach your virtual coffee machine to perform three basic actions: collect the money, replenish the supplies, and serve the coffee.
Program the machine to display how many supplies it has on the screen. Work on the main loop: now, the menu keeps updating until users choose to exit.